@@ -1,16 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Diagnostics.Contracts;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Modular.Core;
 using Modular.Modules.Core.Models;
 
 namespace Modular.Modules.Core.Controllers
 {
 
-
+    [Controller]
     public class HomeController : Controller
     {
         private readonly UserManager<User> _userManager;
@@ -23,8 +26,11 @@ namespace Modular.Modules.Core.Controllers
             _signInManager = signInManager;
 
         }
+
+        //[MinimumAgeAuthorize(10)]
         public IActionResult Index()
         {
+
             return View();
         }
 
