@@ -23,7 +23,7 @@
                             No es un full day más, es <strong>
                                 LA PROPIA EXCURSIÓN
 
-                                @await Component.InvokeAsync("EmployeeList")
+                                @*@await Component.InvokeAkync("EmployeeList")*@
 
                             </strong>.
                         </p>
