@@ -1,6 +1,6 @@
 ﻿@Model TestModels
 <h1> module C</h1>
 
-<h2>Another test: @ViewBag.AnotherTestData @Model.</h2>
+<h2>Another test: @ViewBag.AnotherTestData @Model</h2>
 
 @await Component.InvokeAsync("EmployeeList")
